+++ src/templates/Templates.unit.js
@@ -17,7 +17,7 @@ const componentDir = 'components';
 const fixComponent = (instance, index = 0) => {
   instance.id = instance.key;
   index++;
-  if (instance.type === 'form') {
+  if (instance.everyComponent) {
     instance.everyComponent(component => fixComponent(component, index));
     if (instance.hasOwnProperty('subForm') && instance.subForm) {
       instance.subForm.id = instance.key;